+++ src/gr/aueb/cf/cf9/ch18/bankapp/service/IAccountService.java
@@ -7,12 +7,13 @@ import gr.aueb.cf.cf9.ch18.bankapp.dto.AccountInsertDTO;
 import gr.aueb.cf.cf9.ch18.bankapp.dto.AccountReadOnlyDTO;
 import gr.aueb.cf.cf9.ch18.bankapp.dto.AccountWithdrawDTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface IAccountService {
     AccountReadOnlyDTO createNewAccount(AccountInsertDTO insertDTO);
     void deposit(AccountDepositDTO depositDTO) throws AccountNotFoundException;
     void withdraw(AccountWithdrawDTO withdrawDTO) throws InsufficientBalanceException, AccountNotFoundException;
-    double getBalance(String iban) throws AccountNotFoundException;
+    BigDecimal getBalance(String iban) throws AccountNotFoundException;
     List<AccountReadOnlyDTO> getAllAccounts();
 }
